requirements.txt
antspyx==0.4.2
nibabel==5.2.0
pandas==2.1.4
scipy==1.11.4
SimpleITK==2.3.1
tqdm==4.66.1
